+++ app/javascript/applications/HomebrewsApp/components/substances/Dnd/Feat.jsx
@@ -1,0 +1,51 @@
+import { Show } from 'solid-js';
+
+import { Button, DndFeatForm, createModal } from '../../../components';
+import { useAppLocale } from '../../../context';
+import { Trash, Edit } from '../../../assets';
+
+const TRANSLATION = {
+  en: {
+    level: 'Level'
+  },
+  ru: {
+    level: 'Уровень'
+  }
+}
+
+export const DndFeat = (props) => {
+  const [locale] = useAppLocale();
+  const { Modal, openModal, closeModal } = createModal();
+
+  return (
+    <>
+      <div class="grid grid-cols-12 p-2" classList={{ 'bg-gray-100': props.index && props.index % 2 === 1 }}>
+        <div class="col-span-11">
+          <p class="font-medium!">{props.feature.title.en}</p>
+          <p class="mt-1">{TRANSLATION[locale()].level} {props.feature.conditions.level}</p>
+          <p
+            class="feat-markdown mt-1"
+            innerHTML={props.feature.markdown_description.en} // eslint-disable-line solid/no-innerhtml
+          />
+        </div>
+        <div class="col-span-1 flex items-start justify-end gap-2">
+          <Show when={props.open !== undefined && !props.open}>
+            <Button default classList="p-1" onClick={openModal}>
+              <Edit width="16" height="16" />
+            </Button>
+            <Button default classList="p-1" onClick={() => props.onRemoveFeature(props.feature)}>
+              <Trash width="16" height="16" />
+            </Button>
+          </Show>
+        </div>
+      </div>
+      <Modal>
+        <DndFeatForm
+          feature={props.feature}
+          onSave={(payload) => props.updateFeature(props.feature.id, props.originId, payload)}
+          onCancel={closeModal}
+        />
+      </Modal>
+    </>
+  );
+}