+++ app/javascript/applications/HomebrewsApp/requests/deleteSubclassRequest.jsx
@@ -1,0 +1,8 @@
+import { apiRequest, options } from '../helpers';
+
+export const deleteSubclassRequest = async (accessToken, provider, id) => {
+  return await apiRequest({
+    url: `/homebrews/${provider}/subclasses/${id}.json`,
+    options: options('DELETE', accessToken)
+  });
+}